Using float instead of decimal, otherwise conflict at add to bilaterals

+++ api/management/commands/ingest_appeals.py
@@ -1,6 +1,5 @@
 import os
 import json
-from decimal import Decimal
 from requests import Session, exceptions as reqexc
 from requests.adapters import HTTPAdapter
 from datetime import datetime, timezone, timedelta
@@ -276,7 +275,7 @@ class Command(BaseCommand):
             'amount_requested': detail1['APD_amountCHF'],
             'amount_funded': amount_funded,
             'real_data_update': modify_time,
-            'triggering_amount': Decimal(int(triggering_amount) % 10**10)  # to avoid overflow
+            'triggering_amount': float(int(triggering_amount) % 10**10)  # to avoid overflow
         }
 
         if event is not None:
